feat: switch regret storage to AtomicI64 with ×100k fixed-point scaling
Regrets, predictions, and baselines change from AtomicI32 to AtomicI64
with REGRET_SCALE = 100,000. This restores sub-chip precision lost in
the previous scaling refactor (commit 7815fe8) which truncated fractional
chip deltas to zero.

Memory: regret buffers double from 4 to 8 bytes per slot (~33% total increase
since strategy_sums were already i64). Old i32 snapshots are incompatible.

+++ crates/core/src/cfr/optimizer.rs
@@ -5,10 +5,10 @@
 //! to be selected via config.
 
 // Precision loss on cast to f64 is acceptable for action counts and
-// regret values. Truncation from f64 back to i32/i64 is intentional.
+// regret values. Truncation from f64 back to i64 is intentional.
 #![allow(clippy::cast_possible_truncation, clippy::cast_precision_loss)]
 
-use std::sync::atomic::{AtomicI32, AtomicI64, Ordering};
+use std::sync::atomic::{AtomicI64, Ordering};
 
 use rayon::prelude::*;
 
@@ -25,9 +25,9 @@ pub trait CfrOptimizer: Send + Sync {
     /// DCFR ignores it.
     fn apply_discount(
         &self,
-        regrets: &[AtomicI32],
+        regrets: &[AtomicI64],
         strategy_sums: &[AtomicI64],
-        predictions: Option<&[AtomicI32]>,
+        predictions: Option<&[AtomicI64]>,
         iteration: u64,
     );
 
@@ -39,8 +39,8 @@ pub trait CfrOptimizer: Send + Sync {
     /// `predictions` is the optional prediction buffer.
     fn current_strategy(
         &self,
-        regrets: &[AtomicI32],
-        predictions: Option<&[AtomicI32]>,
+        regrets: &[AtomicI64],
+        predictions: Option<&[AtomicI64]>,
         offset: usize,
         num_actions: usize,
         out: &mut [f64],
@@ -70,9 +70,9 @@ pub struct DcfrOptimizer {
 impl CfrOptimizer for DcfrOptimizer {
     fn apply_discount(
         &self,
-        regrets: &[AtomicI32],
+        regrets: &[AtomicI64],
         strategy_sums: &[AtomicI64],
-        _predictions: Option<&[AtomicI32]>,
+        _predictions: Option<&[AtomicI64]>,
         iteration: u64,
     ) {
         let tf = iteration as f64;
@@ -85,7 +85,7 @@ impl CfrOptimizer for DcfrOptimizer {
         regrets.par_iter().for_each(|atom| {
             let v = atom.load(Ordering::Relaxed);
             let d = if v >= 0 { d_pos } else { d_neg };
-            atom.store((f64::from(v) * d) as i32, Ordering::Relaxed);
+            atom.store((v as f64 * d) as i64, Ordering::Relaxed);
         });
 
         strategy_sums.par_iter().for_each(|atom| {
@@ -96,8 +96,8 @@ impl CfrOptimizer for DcfrOptimizer {
 
     fn current_strategy(
         &self,
-        regrets: &[AtomicI32],
-        _predictions: Option<&[AtomicI32]>,
+        regrets: &[AtomicI64],
+        _predictions: Option<&[AtomicI64]>,
         offset: usize,
         num_actions: usize,
         out: &mut [f64],
@@ -106,7 +106,7 @@ impl CfrOptimizer for DcfrOptimizer {
         let mut sum = 0.0_f64;
         for a in 0..num_actions {
             let r = regrets[offset + a].load(Ordering::Relaxed);
-            let v = if r > 0 { f64::from(r) } else { 0.0 };
+            let v = if r > 0 { r as f64 } else { 0.0 };
             out[a] = v;
             sum += v;
         }
@@ -149,9 +149,9 @@ pub struct SapcfrPlusOptimizer {
 impl CfrOptimizer for SapcfrPlusOptimizer {
     fn apply_discount(
         &self,
-        regrets: &[AtomicI32],
+        regrets: &[AtomicI64],
         strategy_sums: &[AtomicI64],
-        _predictions: Option<&[AtomicI32]>,
+        _predictions: Option<&[AtomicI64]>,
         iteration: u64,
     ) {
         let tf = iteration as f64;
@@ -162,7 +162,7 @@ impl CfrOptimizer for SapcfrPlusOptimizer {
         // Discount regrets: floor negative to 0 (RM+ style).
         regrets.par_iter().for_each(|atom| {
             let v = atom.load(Ordering::Relaxed);
-            let discounted = (f64::from(v) * d_pos) as i32;
+            let discounted = (v as f64 * d_pos) as i64;
             atom.store(discounted.max(0), Ordering::Relaxed);
         });
 
@@ -175,8 +175,8 @@ impl CfrOptimizer for SapcfrPlusOptimizer {
 
     fn current_strategy(
         &self,
-        regrets: &[AtomicI32],
-        predictions: Option<&[AtomicI32]>,
+        regrets: &[AtomicI64],
+        predictions: Option<&[AtomicI64]>,
         offset: usize,
         num_actions: usize,
         out: &mut [f64],
@@ -186,7 +186,7 @@ impl CfrOptimizer for SapcfrPlusOptimizer {
         for a in 0..num_actions {
             let r = regrets[offset + a].load(Ordering::Relaxed);
             let v = predictions.map_or(0, |p| p[offset + a].load(Ordering::Relaxed));
-            let predicted = f64::from(r) + self.eta * f64::from(v);
+            let predicted = r as f64 + self.eta * v as f64;
             let clamped = predicted.max(0.0);
             out[a] = clamped;
             sum += clamped;
@@ -214,8 +214,6 @@ impl CfrOptimizer for SapcfrPlusOptimizer {
 
 #[cfg(test)]
 mod tests {
-    use std::sync::atomic::{AtomicI32, AtomicI64, Ordering};
-
     // Tests for DcfrOptimizer
 
     #[test]
@@ -226,7 +224,7 @@ mod tests {
             beta: 0.0,
             gamma: 2.0,
         };
-        let regrets = vec![AtomicI32::new(1000), AtomicI32::new(-500)];
+        let regrets = vec![AtomicI64::new(1000), AtomicI64::new(-500)];
         let strats = vec![AtomicI64::new(2000)];
         opt.apply_discount(&regrets, &strats, None, 10);
 
@@ -250,7 +248,7 @@ mod tests {
             beta: 0.0,
             gamma: 2.0,
         };
-        let regrets: Vec<AtomicI32> = vec![];
+        let regrets: Vec<AtomicI64> = vec![];
         let strats = vec![AtomicI64::new(2000)];
         opt.apply_discount(&regrets, &strats, None, 10);
 
@@ -268,9 +266,9 @@ mod tests {
             beta: 0.0,
             gamma: 2.0,
         };
-        let regrets = vec![AtomicI32::new(1000)];
+        let regrets = vec![AtomicI64::new(1000)];
         let strats = vec![AtomicI64::new(2000)];
-        let preds = vec![AtomicI32::new(999)];
+        let preds = vec![AtomicI64::new(999)];
         // Passing predictions should not change behavior for DCFR
         opt.apply_discount(&regrets, &strats, Some(&preds), 10);
         let r0 = regrets[0].load(Ordering::Relaxed);
@@ -286,9 +284,9 @@ mod tests {
             gamma: 2.0,
         };
         let regrets = vec![
-            AtomicI32::new(300),  // action 0
-            AtomicI32::new(100),  // action 1
-            AtomicI32::new(-50),  // action 2 (negative, excluded)
+            AtomicI64::new(300),  // action 0
+            AtomicI64::new(100),  // action 1
+            AtomicI64::new(-50),  // action 2 (negative, excluded)
         ];
         let mut out = [0.0; 3];
         opt.current_strategy(&regrets, None, 0, 3, &mut out);
@@ -307,8 +305,8 @@ mod tests {
             gamma: 2.0,
         };
         let regrets = vec![
-            AtomicI32::new(-100),
-            AtomicI32::new(-200),
+            AtomicI64::new(-100),
+            AtomicI64::new(-200),
         ];
         let mut out = [0.0; 2];
         opt.current_strategy(&regrets, None, 0, 2, &mut out);
@@ -326,10 +324,10 @@ mod tests {
         };
         // Simulate a buffer with multiple nodes. Node at offset 2 has 2 actions.
         let regrets = vec![
-            AtomicI32::new(0),    // slot 0 (different node)
-            AtomicI32::new(0),    // slot 1 (different node)
-            AtomicI32::new(600),  // slot 2 = our node, action 0
-            AtomicI32::new(200),  // slot 3 = our node, action 1
+            AtomicI64::new(0),    // slot 0 (different node)
+            AtomicI64::new(0),    // slot 1 (different node)
+            AtomicI64::new(600),  // slot 2 = our node, action 0
+            AtomicI64::new(200),  // slot 3 = our node, action 1
         ];
         let mut out = [0.0; 2];
         opt.current_strategy(&regrets, None, 2, 2, &mut out);
@@ -370,8 +368,8 @@ mod tests {
             gamma: 2.0,
             eta: 0.5,
         };
-        let regrets = vec![AtomicI32::new(200), AtomicI32::new(100), AtomicI32::new(0)];
-        let preds = vec![AtomicI32::new(100), AtomicI32::new(-50), AtomicI32::new(50)];
+        let regrets = vec![AtomicI64::new(200), AtomicI64::new(100), AtomicI64::new(0)];
+        let preds = vec![AtomicI64::new(100), AtomicI64::new(-50), AtomicI64::new(50)];
         let mut out = [0.0; 3];
         opt.current_strategy(&regrets, Some(&preds), 0, 3, &mut out);
         // R_tilde[0] = max(0, 200 + 0.5*100) = 250
@@ -407,7 +405,7 @@ mod tests {
             eta: 0.5,
         };
         // Without predictions, eta * 0 = 0 => standard regret matching
-        let regrets = vec![AtomicI32::new(300), AtomicI32::new(100)];
+        let regrets = vec![AtomicI64::new(300), AtomicI64::new(100)];
         let mut out = [0.0; 2];
         opt.current_strategy(&regrets, None, 0, 2, &mut out);
         assert!((out[0] - 0.75).abs() < 0.01);
@@ -423,8 +421,8 @@ mod tests {
             eta: 0.5,
         };
         // Even with predictions, if all predicted regrets are negative => uniform
-        let regrets = vec![AtomicI32::new(-100), AtomicI32::new(-200)];
-        let preds = vec![AtomicI32::new(-100), AtomicI32::new(-100)];
+        let regrets = vec![AtomicI64::new(-100), AtomicI64::new(-200)];
+        let preds = vec![AtomicI64::new(-100), AtomicI64::new(-100)];
         let mut out = [0.0; 2];
         opt.current_strategy(&regrets, Some(&preds), 0, 2, &mut out);
         assert!((out[0] - 0.5).abs() < 0.01);
@@ -439,7 +437,7 @@ mod tests {
             gamma: 2.0,
             eta: 0.5,
         };
-        let regrets = vec![AtomicI32::new(1000), AtomicI32::new(-500)];
+        let regrets = vec![AtomicI64::new(1000), AtomicI64::new(-500)];
         let strats = vec![AtomicI64::new(2000)];
         opt.apply_discount(&regrets, &strats, None, 10);
         // Negative regrets are floored to 0 (RM+ style).
@@ -458,7 +456,7 @@ mod tests {
             gamma: 2.0,
             eta: 0.5,
         };
-        let regrets: Vec<AtomicI32> = vec![];
+        let regrets: Vec<AtomicI64> = vec![];
         let strats = vec![AtomicI64::new(2000)];
         opt.apply_discount(&regrets, &strats, None, 10);
         let s0 = strats[0].load(Ordering::Relaxed);
@@ -498,16 +496,16 @@ mod tests {
         };
         // 4-slot buffer, our node starts at offset 2 with 2 actions
         let regrets = vec![
-            AtomicI32::new(0),
-            AtomicI32::new(0),
-            AtomicI32::new(100), // our action 0
-            AtomicI32::new(100), // our action 1
+            AtomicI64::new(0),
+            AtomicI64::new(0),
+            AtomicI64::new(100), // our action 0
+            AtomicI64::new(100), // our action 1
         ];
         let preds = vec![
-            AtomicI32::new(0),
-            AtomicI32::new(0),
-            AtomicI32::new(200), // prediction for action 0
-            AtomicI32::new(-200), // prediction for action 1 (negative)
+            AtomicI64::new(0),
+            AtomicI64::new(0),
+            AtomicI64::new(200), // prediction for action 0
+            AtomicI64::new(-200), // prediction for action 1 (negative)
         ];
         let mut out = [0.0; 2];
         opt.current_strategy(&regrets, Some(&preds), 2, 2, &mut out);
